+++ spark/src/main/scala-shims/spark-4.2/VariantShreddingShims.scala
@@ -16,21 +16,57 @@
 
 package org.apache.spark.sql.delta.shims
 
+import scala.util.control.NonFatal
+
+import org.apache.hadoop.conf.Configuration
+import org.apache.hadoop.fs.Path
+import org.apache.parquet.format.converter.ParquetMetadataConverter
+import org.apache.parquet.hadoop.util.HadoopInputFile
+
+import org.apache.spark.sql.delta.stats.{DeltaTaskStatisticsTracker, VariantStatsData}
+import org.apache.spark.sql.execution.datasources.{DataSourceUtils, OutputWriter, WriteTaskStatsTracker}
+import org.apache.spark.sql.execution.datasources.parquet.ParquetFooterReader
 import org.apache.spark.sql.internal.SQLConf
 
-/**
- * Shim for variant shredding configs to handle API changes between Spark versions.
- * In Spark 4.2, VARIANT_INFER_SHREDDING_SCHEMA config exists.
- *
- * This shim provides a way to conditionally add the config to the options map
- * when writing files.
- */
 object VariantShreddingShims {
-  /**
-   * Returns a Map containing variant shredding related configs for file writing.
-   * In Spark 4.2, this returns the VARIANT_INFER_SHREDDING_SCHEMA config.
-   */
   def getVariantInferShreddingSchemaOptions(enableVariantShredding: Boolean): Map[String, String] = {
     Map(SQLConf.VARIANT_INFER_SHREDDING_SCHEMA.key -> enableVariantShredding.toString)
+  }
+
+  def extractAndInjectVariantStats(
+      writer: OutputWriter,
+      trackers: Seq[WriteTaskStatsTracker],
+      parquetRebaseModeInRead: String,
+      hadoopConf: Configuration): Unit = {
+    val hasVariantTrackers = trackers.exists {
+      case t: DeltaTaskStatisticsTracker =>
+        t.minVariantStatsExpressions.nonEmpty || t.maxVariantStatsExpressions.nonEmpty
+      case _ => false
+    }
+    if (!hasVariantTrackers) return
+    val footer = try {
+      ParquetFooterReader.readFooter(
+        HadoopInputFile.fromPath(new Path(writer.path()), hadoopConf),
+        ParquetMetadataConverter.NO_FILTER)
+    } catch {
+      case NonFatal(_) => return
+    }
+    val keyValueMetadata = footer.getFileMetaData.getKeyValueMetaData
+    val rebaseSpec = DataSourceUtils.datetimeRebaseSpec(
+      k => keyValueMetadata.get(k),
+      parquetRebaseModeInRead)
+    val dateRebaseFunc =
+      DataSourceUtils.createDateRebaseFuncInRead(rebaseSpec.mode, "Parquet")
+    val timestampRebaseFunc =
+      DataSourceUtils.createTimestampRebaseFuncInRead(rebaseSpec, "Parquet")
+    val variantStatsData = VariantStatsData(footer, dateRebaseFunc, timestampRebaseFunc)
+    trackers.foreach {
+      case tracker: DeltaTaskStatisticsTracker =>
+        tracker.minVariantStatsExpressions.foreach(
+          _.addVariantStatsData(variantStatsData))
+        tracker.maxVariantStatsExpressions.foreach(
+          _.addVariantStatsData(variantStatsData))
+      case _ =>
+    }
   }
 }
