+++ spark/src/main/scala/org/apache/spark/sql/delta/stats/VariantStats.scala
